requirements.txt
gspread==5.12.0
pandas>=1.5.0
google-auth==2.23.4
flask==3.0.0
matplotlib>=3.5.0
seaborn>=0.12.0 

